+++ apps/s3api/src/error.rs
