+++ libs/aws/sabi-s3/src/operations/put_object/file.rs
@@ -1,12 +1,12 @@
-use crate::errors::Error::{StdIoError, PutObjectError};
+use super::error::Error::FileNotFound;
+use crate::error::Error::{PutObjectError, StdIoError};
 use crate::internal::RequestResource;
 use crate::verbs::HasObjectKey;
 use crate::S3Result;
 use reqwest::blocking::Body;
 use std::error::Error;
 use std::fs::File;
-use super::error::Error::FileNotFound;
-use std::io::ErrorKind;
+use std::io::ErrorKind::NotFound;
 
 #[derive(Debug)]
 pub struct FileRequest {
@@ -17,14 +17,11 @@ pub struct FileRequest {
 impl FileRequest {
     fn open_file(&self) -> S3Result<File> {
         File::open(&self.file_path).map_err(|e| match e {
-            _ if e.kind() == ErrorKind::NotFound => {
-                PutObjectError(FileNotFound {
-                    path: self.file_path.to_string(),
-                    description: e.description().to_string(),
-                })
-            }
-            _ => StdIoError(e)
-            ,
+            _ if e.kind() == NotFound => PutObjectError(FileNotFound {
+                path: self.file_path.to_string(),
+                description: e.description().to_string(),
+            }),
+            _ => StdIoError(e),
         })
     }
 }
